+++ CHANGELOG.md
@@ -11,4 +11,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 ### Changed
+- Add Task.Supervisor for async loading tasks (#7)
 - Pre-calculate linear_counts to avoid 4x iteration per render (#4)